chore(ci): fix CI failures so deploy job can run
- pnpm/action-setup: drop pinned version, defer to packageManager
  in package.json (otherwise the action errors on duplicate spec)
- terrain-decoder: drop redundant `as u32` casts (decoder.dimensions
  already returns u32, clippy 1.95 catches it)
- mesh::build_mesh: allow clippy::too_many_arguments
- cargo fmt across height-optimizer and buildings-core

+++ crates/terrain-decoder/src/lib.rs
@@ -58,8 +58,8 @@ pub fn decode_terrarium_webp(z: u8, x: u32, y: u32, bytes: &[u8]) -> Result<Terr
     }
 
     Ok(TerrainTile {
-        width: w as u32,
-        height: h as u32,
+        width: w,
+        height: h,
         z,
         x,
         y,
